hrt: Fix possible memory leak when returning early
